refactor(http): decompose HttpFencedPanel into siblings + a state hook
HttpFencedPanel was a 2634-line component holding toolbar, form/raw
mode, result tabs, status bar, settings drawer, codegen, history,
cache hydration, and the run state machine. Brought under the
600-line size gate by extracting cohesive units.

View siblings (each under 600 L, extracted verbatim):
- `HttpJsonVisualizer.tsx` — JSON tree + flattener for the visualize
  tab.
- `HttpBodyView.tsx` — read-only CM6 body viewer + image/PDF/HTML
  preview overlay + pretty/raw/visualize toggle.
- `HttpInlineEditors.tsx` — the 3 inline CM editors used by the
  form-table cells.
- `HttpFormTables.tsx` — body-tab dispatcher + FormUrlEncodedTable +
  MultipartTable + BinaryFilePicker.

Shared state machine:
- `useExecutableBlock<T>` — the `idle -> running -> success | error |
  cancelled` machine + `AbortController` + collect-blocks/env +
  try/catch/finally. HTTP consumes it through an adapter (`validate`,
  `prepare`, `execute`, `elapsedOf`, `persist?`, `onOutcome?`,
  `onRunStart?`, `onProgress?`). DbFencedPanel deliberately does not
  consume it: its `runExplain` + `loadMore` co-mutate the same state
  as `run`, so migrating would force exposing raw setters and lose
  the encapsulation the hook exists to provide.

Pure helpers:
- `http-request-builder.ts` — `parseBody` / `deriveHost` /
  `httpElapsedOf` / `isValidHeaderName` / `HTTP_TOKEN_RE` /
  `buildExecutorParams`, with full module test coverage.

Cohesive concerns split into 4 sibling hooks:
- `useHttpRefsContext.ts` — refs context + stable getters.
- `useHttpCacheHydrate.ts` — cache lookup on mount/body change;
  mutations skipped.
- `useHttpCodegenSnippets.ts` — cURL/fetch/Python/HTTPie/.http
  pre-computation + `handleSendAs` + `copyAsCurl`.
- `useHttpDrawerData.ts` — history + examples loaders + recordHistory
  + drawer actions.

Result: HttpFencedPanel 2634 to 567 L. It now fits under the size
gate with no escape-hatch pragma.
